Add files via upload
Must read...
If I have learned anything in this long life of mine, it is this: In love we find out who we want to be; in
war we find out who we are. Today’s young people want to know everything about everyone. They
think talking about a problem will solve it. I come from a quieter generation. We understand the value
of forgetting, the lure of reinvention.